requirements.txt
aiogram==3.7.0
aiosqlite
python-dotenv
aiohttp
